дополнительная декомпозиция (добавлены блок с комментариями на странице предложения, форма логина, рефакторинг кнопки для избранного)
