+++ src/pages/login-page/login-page.tsx
@@ -1,33 +1,9 @@
 import {Helmet} from 'react-helmet-async';
 import Logo from '../../components/logo/logo';
-import {useRef, FormEvent} from 'react';
-import {useNavigate} from 'react-router-dom';
-import {useAppDispatch} from '../../hooks/index';
-import {loginAction} from '../../store/api-actions';
-import {AppRoute} from '../../consts';
+
+import LoginForm from '../../components/login-form/login-form';
 
 function LoginPage(): JSX.Element {
-  const loginRef = useRef<HTMLInputElement | null>(null);
-  const passwordRef = useRef<HTMLInputElement | null>(null);
-
-  const dispatch = useAppDispatch();
-  const navigate = useNavigate();
-
-  const handleLoginFormSubmit = (evt: FormEvent<HTMLFormElement>) => {
-    evt.preventDefault();
-
-    if (loginRef.current !== null && passwordRef.current !== null) {
-      dispatch(loginAction({
-        email: loginRef.current.value,
-        password: passwordRef.current.value
-      }))
-        .then((response) => {
-          if (response.meta.requestStatus === 'fulfilled') {
-            navigate(AppRoute.Main);
-          }
-        });
-    }
-  };
 
   return (
     <div className="page page--gray page--login">
@@ -47,44 +23,7 @@ function LoginPage(): JSX.Element {
         <div className="page__login-container container">
           <section className="login">
             <h1 className="login__title">Sign in</h1>
-            <form
-              className="login__form form"
-              action="#"
-              method="post"
-              onSubmit={handleLoginFormSubmit}
-            >
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">E-mail</label>
-                <input
-                  ref={loginRef}
-                  className="login__input form__input"
-                  type="email"
-                  name="email"
-                  placeholder="Email"
-                  required
-                />
-              </div>
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">Password</label>
-                <input
-                  ref={passwordRef}
-                  className="login__input form__input"
-                  type="password"
-                  name="password"
-                  placeholder="Password"
-                  minLength={2}
-                  pattern='^.*(?=.*[a-zA-Z])(?=.*\d).*$'
-                  title='Пароль должен содержать буквы латинского алфавита и цифры: минимум 1 букву и 1 цифру'
-                  required
-                />
-              </div>
-              <button
-                className="login__submit form__submit button"
-                type="submit"
-              >
-            Sign in
-              </button>
-            </form>
+            <LoginForm/>
           </section>
           <section className="locations locations--login locations--current">
             <div className="locations__item">
