+++ src/pages/offer-page/offer-page.tsx
@@ -7,27 +7,22 @@ import OfferGallery from '../../components/offer-gallery/offer-gallery';
 import Rating from '../../components/rating/rating';
 import OfferHost from '../../components/offer-host/offer-host';
 import OffersList from '../../components/offers-list/offers-list';
-import ReviewForm from '../../components/review-form/review-form';
-import ReviewsList from '../../components/reviews-list/reviews-list';
+import ReviewsContainer from '../../components/reviews-container/reviews-container';
 import OfferInsideList from '../../components/offer-inside-list/offer-inside-list';
 import Map from '../../components/map/map';
 import NotFoundPage from '../../pages/not-found-page/not-found-page';
-import {CardType, MapTypes, AuthorizationStatus, OfferCardCount} from '../../consts';
+import {CardType, MapTypes, OfferCardCount, FavouriteButtonType} from '../../consts';
 import LoadingPage from '../../pages/loading-page/loading-page';
 import {useAppDispatch, useAppSelector} from '../../hooks/index';
 import {fetchOfferDataAction, fetchOfferReviewsAction, fetchNearbyPlacesAction} from '../../store/api-actions';
 import {capitalize, getMapPoints} from '../../utils/common';
 import {getCurrentCity} from '../../store/app-process-slice/selectors';
-import {getAuthorizationStatus} from '../../store/user-process-slice/selectors';
 import {getFullOfferLoadingStatus, getFullOfferData, getNearbyPlaces} from '../../store/full-offer-process-slice/selectors';
-import {getReviews} from '../../store/review-process-slice/selectors';
 
 function OfferPage(): JSX.Element {
   const currentCity = useAppSelector(getCurrentCity);
-  const currentAuthorizationStatus = useAppSelector(getAuthorizationStatus);
   const isFullOfferLoading = useAppSelector(getFullOfferLoadingStatus);
   const currentOfferData = useAppSelector(getFullOfferData);
-  const reviews = useAppSelector(getReviews);
   const nearbyPlaces = useAppSelector(getNearbyPlaces).slice(OfferCardCount.Min, OfferCardCount.Max);
 
   const params = useParams();
@@ -71,7 +66,7 @@ function OfferPage(): JSX.Element {
                 <h1 className="offer__name">
                   {title}
                 </h1>
-                <FavoriteButton buttonType={'activeOfferButton'} isFavorite={isFavorite} />
+                <FavoriteButton buttonType={FavouriteButtonType.FullOfferButton} isFavorite={isFavorite} />
               </div>
               <Rating ratingType={'averageOfferRating'} ratingValue={rating}/>
               <ul className="offer__features">
@@ -97,14 +92,7 @@ function OfferPage(): JSX.Element {
                   </p>
                 </div>
               </div>
-              <section className="offer__reviews reviews">
-                {reviews.length > 0 ? (
-                  <h2 className="reviews__title">
-              Reviews · <span className="reviews__amount">{reviews.length}</span>
-                  </h2>) : ''}
-                <ReviewsList reviews={reviews} />
-                {currentAuthorizationStatus === AuthorizationStatus.Auth && <ReviewForm />}
-              </section>
+              <ReviewsContainer/>
             </div>
           </div>
           <Map mapPoints={mapPoints} cityLocation={currentCity.location} activeOffer={activeOfferId} mapType={MapTypes.Offer}/>
